Initial develoment for OpenBSD 6.1 + Collectd
Preparations to upgrade from OpenBSD 6 to 6.1.
Installing and configuring Collectd.
Installing and configuring rrdscout.
Installing and configuring MariaDB (to allow smoking of DBD::Mysql).
Changed the way CPAN preference files (YAML) are stored for smoker
users: now all files are shared from /minicpan directory. That will make
it easier to create new files and share them between the users.
Enabling a "fake" local proxy for smoker users to avoid 404 errors (that
take a long time to be resolved since CPAN tries to fetch those files
with different procedures).
Removed automatic usage of mirror_cleanup: still to be validate if it is
not causing 404 errors of cpan client.