Improvements
The bootstrap for vagrant user wasn't using full paths, and maybe that's
the reason for failing to setup .bash_profile to setup perlbrew at logon
time.

Sharing the code related to cleanup_cpan wasn't working as expected,
took a step back to copy and paste the single line of code that was
doing this job to both vagrant_user and vagrant_bootstrap scripts.