New packages
Added gnupg2 so modules that use it for testing can make it so.
llvm is the C/C++ compiler of choice of OpenBSD and has more extented
support for C++ than the GNU compiler g++.
We want to use the standard perl that comes with OpenBSD, so local::lib
is required as well.